[Fix](Langmapper): Reverted hnd to en mapping, now configure everything in hnd.
