+++ lua/ds_omega/config/EditorManagement/langmapper.lua
@@ -10,31 +10,24 @@ return {
           return output[#output]
         end
       end
-
-      local ru = [[ёйцукенгшщзхъфывапролджэ\\ячсмитьбю.]]
-      local en = [[`qwertyuiop[]asdfghjkl;'\\zxcvbnm,./]]
-      local hands_down_neu = [[`wfmpv/.q"'z(rsntg,aeihj)\xcldb-uoyk]]
-
-      local ru_shift = [[ЁЙЦУКЕНГШЩЗХЪФЫВАПРОЛДЖЭ//ЯЧСМИТЬБЮ,]]
-      local en_shift = [[~QWERTYUIOP{}ASDFGHJKL:"||ZXCVBNM<>?]]
-      local hands_down_neu_shift = [[~WFMPV*:Q[]Z{RSNTG;AEIHJ}|XCLDB+UOYK]]
+      local layouts = require('langmap').layouts
 
       return {
           ---@type boolean Wrap all keymap's functions (nvim_set_keymap etc)
           hack_keymap = false,
           ---@type string Standart English layout (on Mac, It may be different in your case.)
-          default_layout = en .. en_shift,
+          default_layout = layouts.en .. layouts.en_shift,
           ---@type string[] Names of layouts. If empty, will handle all configured layouts.
-          use_layouts = { 'hdn', 'ru' },
+          use_layouts = { 'ru' },
           layouts = {
               ru = {
-                  id = '1049',
-                  layout = ru .. ru_shift,
-                      default_layout = nil,
+                id = '1049',
+                layout = layouts.ru .. layouts.ru_shift,
+                default_layout = nil,
               },
               hdn = {
                   id = '1033',
-                  layout = hands_down_neu .. hands_down_neu_shift,
+                  layout = layouts.hands_down_neu .. layouts.hands_down_neu_shift,
                   -- layout = [[wfmpv/.q"'z(rsntg,aeihj)xcldb-uoykWFMPV*:Q[]Z{RSNTG;AEIHJ}XCLDB+UOYK]],
                   -- default_layout = [[qwertyuiop[]asdfghjkl;'\zxcvbnm,./QWERTYUIOP{}ASDFGHJKL:"|ZXCVBNM<>?]],
                   default_layout = nil,
